Remove mysqldump flags
Remove mysqldump flags

Hmm ok so I have confused myself with this...The docker container we are using to run the tests is using mariadb-client and not mysql-client  https://github.com/alphagov/govwifi-database-backup/blob/master/Dockerfile#L8. I assumed we were using the mysql-client version of mysqldump when I added the flags to mysqldump in the previous commits:

https://github.com/alphagov/govwifi-database-backup/pull/20
https://github.com/alphagov/govwifi-database-backup/pull/18

This difference is important, because the mariadb version of mysqldump accepts different flags. Turns out the two flags I added previously don’t exist in this version.