Fix Error With Github Actions Tests
Fix error with github actions tests (unit tests). The tests were failing
because the dummy s3 bucket was not being created. The script that creates the
bucket was never being called. Added a line that runs the command for creating
the local dummy S3 bucket (used for simulating how the script will behave in
AWS).

Removed "docker-compose build" command, this is not needed and was causing
another error. Replaced with "docker-compose up --build" this will build the
images fresh each time AND bring the containers up.

Removed an unnecessary volume from the "fake-s3" container, and an unnecessary
directory and script. This script was never called.

Added extra explanation to Readme about the different roles of the Dockerfile
vs docker-compose.yml files